Moved groupBy to /utilities/array.ts
