+++ packages/apollo-gateway/src/utilities/array.ts
@@ -48,3 +48,22 @@ export function findAndExtract<T>(
 
   return [array[index], remaining];
 }
+
+export function groupBy<T, U>(keyFunction: (element: T) => U) {
+  return (iterable: Iterable<T>) => {
+    const result = new Map<U, T[]>();
+
+    for (const element of iterable) {
+      const key = keyFunction(element);
+      const group = result.get(key);
+
+      if (group) {
+        group.push(element);
+      } else {
+        result.set(key, [element]);
+      }
+    }
+
+    return result;
+  };
+}
